menue.h: fix type conversion
Wconversion

+++ src/gui/widget/menue.h
@@ -641,7 +641,7 @@ class CMenuWidget : public CMenuTarget, public CComponentsSignals
 		void removeItem(const uint& item_id);
 		int getItemId(CMenuItem *menuItem);
 		int getItemId(const char *name);
-		int getItemsCount()const{return items.size();};
+		int getItemsCount()const{return (int)items.size();};
 		CMenuItem* getItem(const uint& item_id);
 		virtual void paint();
 		virtual void hide();
